+++ CHANGELOG.md
@@ -36,6 +36,7 @@ And please only add new entries to the top of this list, right below the `# Unre
 - On Wayland, fix `TouchPhase::Ended` always reporting the location of the first touch down, unless the compositor
   sent a cancel or frame event.
 - On iOS, send `RedrawEventsCleared` even if there are no redraw events, consistent with other platforms.
+- On MacOS implementation left and right option keys are detected as `LAlt` and `RAlt` modifiers.
 
 # 0.26.1 (2022-01-05)